+++ ChangeLog.md
@@ -1,3 +1,9 @@
+2.19.0 [2020-07-29]
+
+  * change `smof uniq --final-header` to `--first-header` and keep first header
+    rather than last
+  * add `smof uniq --removed=FILE` option that writes removed entries to FILE
+
 2.18.0 [2020-03-02]
 
   * allow lowercase 'X' to be replace with 'n' in clean standard function